tools/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mantools"
version = "0.1.0"
description = "GitOps AI Agent — Multi-SCM autonomous git operations powered by LLM"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "mamatnurahmat", email = "mamatnurahmat@gmail.com"},
]
keywords = ["gitops", "ai-agent", "llm", "devops", "git", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Version Control :: Git",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]
dev = [
    "build",
    "twine",
    "pytest",
]

[project.scripts]
mantools = "agents.cli:main"

[project.urls]
Homepage = "https://github.com/mamatnurahmat/mantools"
Repository = "https://github.com/mamatnurahmat/mantools"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*"]
